+++ recipiece_backend/src/api/recipe/share/listRecipeShares.ts
@@ -1,9 +1,9 @@
 import { Prisma } from "@prisma/client";
-import { ListRecipeSharesQuerySchema, ListRecipeSharesResponseSchema } from "../../schema";
-import { ApiResponse, AuthenticatedRequest } from "../../types";
+import { ListRecipeSharesQuerySchema, ListRecipeSharesResponseSchema } from "../../../schema";
+import { ApiResponse, AuthenticatedRequest } from "../../../types";
 import { StatusCodes } from "http-status-codes";
-import { DEFAULT_PAGE_SIZE } from "../../util/constant";
-import { prisma } from "../../database";
+import { DEFAULT_PAGE_SIZE } from "../../../util/constant";
+import { prisma } from "../../../database";
 
 /**
  * List recipes shares that are targeting the requesting user or the requesting user has sent.
